api/sentiment/requirements.txt
adapter-transformers==2.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
filelock==3.3.2
Flask==2.0.2
fugashi==1.1.1
huggingface-hub==0.0.19
idna==3.3
ipadic==1.0.0
itsdangerous==2.0.1
Jinja2==3.0.2
joblib==1.1.0
MarkupSafe==2.0.1
numpy==1.21.3
packaging==21.2
pyparsing==2.4.7
PyYAML==6.0
regex==2021.10.23
requests==2.26.0
sacremoses==0.0.46
six==1.16.0
tokenizers==0.10.3
torch==1.10.0
tqdm==4.62.3
typing-extensions==3.10.0.2
unidic-lite==1.0.8
urllib3==1.26.7
Werkzeug==2.0.2
